requirements.txt
langchain
openai
faiss-cpu
streamlit
pymupdf
python-dotenv